Move App component into the components folder together with global CSS
