+++ src/components/App/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { Route, Switch } from 'react-router-dom'
-import Home from './pages/Home'
+import Home from '../../pages/Home'
+import './index.css'
 
 const App: React.FC = () => {
 	return (
