+++ src/index.tsx
@@ -1,8 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { BrowserRouter } from 'react-router-dom'
-import './index.css'
-import App from './App'
+import App from './components/App'
 import reportWebVitals from './utils/reportWebVitals'
 
 ReactDOM.render(
